% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwards-package.r
\docType{package}
\name{forwards}
\alias{forwards}
\alias{forwards-package}
\title{Data Released by Forwards}
\description{
\code{forwards} provides data sets released by \href{Forwards}{http://forwards.github.io/}, the R Foundation taskforce on women and other under-represented groups.
}
